add Boost's chrono library instead of date_time as a dependency when building OpenImageIO
OpenImageIO does not actually require the Boost::date_time library.
More detail is available here:

https://github.com/OpenImageIO/oiio/pull/2727

However, when building on Mac, specifying that Boost::thread should be
installed seems not to also cause Boost::chrono to be installed,
which is a dependency for Boost::thread. We explicitly install
Boost::chrono when building OpenImageIO to ensure that the dependency
for Boost::thread is in place.