add a mechanism for downloaders to save files with a prefix
As of this writing, Imath and OpenEXR are dangerously close to sharing the same
version number. In that case, build_usd.py would download Imath and install it,
then later check whether it needs to download OpenEXR and opt not to since
there would already be a file in place that matches the expected name.

This allows a filename prefix to be specified for a downloader, and it is put
into use for Imath and OpenEXR for now to avoid this potential filename
collision.