bump Boost to version 1.85.0 for Python 3.11+ and on MacOS
Boost 1.85.0 is the version specified by the VFX reference platform for CY 2025.

CMake supports Boost 1.85 as of version 3.29.3, but Visual Studio 2022 does not
yet offer a release bundling a 3.29.X version of CMake. Using the Boost option
"Boost_ADDITIONAL_VERSIONS=1.85" is needed to support building OpenUSD using
the version of CMake currently bundled with the latest release of Visual Studio
2022.