build_usd.py: patch libjpeg-turbo to allow building macOS universal binaries
libjpeg-turbo uses assembly code when SIMD extensions are enabled,
in which case it does not support building universal binaries for
macOS in a single invocation of CMake.

To make building macOS universal binaries more convenient with
build_usd.py, we disable libjpeg-turbo SIMD extensions and patch
its CMake setup to allow the build to proceed, even though we may
be sacrificing performance.